move bench code
