+++ fastrace/Cargo.toml
@@ -54,10 +54,6 @@ name = "trace"
 
 [[bench]]
 harness = false
-name = "compare"
-
-[[bench]]
-harness = false
 name = "spsc"
 
 [profile.release]
